+++ apps/dine-pleiepenger/src/components/error-boundary/ErrorBoundary.tsx
@@ -1,4 +1,4 @@
-import { ContentContainer } from '@navikt/ds-react';
+import { Page } from '@navikt/ds-react';
 import { Component, ErrorInfo, PropsWithChildren, ReactNode } from 'react';
 import { createChildLogger } from '@navikt/next-logger';
 import { getUserRequestId } from '../../utils/userRequestId';
@@ -35,9 +35,11 @@ class ErrorBoundary extends Component<PropsWithChildren, State> {
     render(): ReactNode {
         if (this.state.hasError) {
             return (
-                <ContentContainer>
-                    <ErrorFallback />
-                </ContentContainer>
+                <Page>
+                    <Page.Block width="xl">
+                        <ErrorFallback />
+                    </Page.Block>
+                </Page>
             );
         }