+++ packages/sif-common-core-ds/src/components/progress-stepper/ProgressStepper.tsx
@@ -1,5 +1,5 @@
 import { BodyShort, Heading, Stepper } from '@navikt/ds-react';
-import Step, { StepperStepProps } from '@navikt/ds-react/esm/stepper/Step';
+import { StepperStepProps } from '@navikt/ds-react/Stepper';
 import React, { useEffect, useRef, useState } from 'react';
 import { Back, Collapse, Expand } from '@navikt/ds-icons';
 import { guid } from '@navikt/sif-common-utils';
@@ -148,13 +148,13 @@ const ProgressStepper: React.FunctionComponent<Props> = ({
                                 activeStep={currentStepNumber}
                                 onStepChange={onStepSelect ? handleStepChange : undefined}>
                                 {steps.map((s) => (
-                                    <Step
+                                    <Stepper.Step
                                         href="#"
                                         key={s.id}
                                         completed={s.completed}
                                         interactive={onStepSelect !== undefined && s.completed === true}>
                                         {s.label}
-                                    </Step>
+                                    </Stepper.Step>
                                 ))}
                             </Stepper>
                             {allStepsFooter && (