+++ packages/sif-common-formik-ds/storybook/components/validation-panel/ValidationPanel.tsx
@@ -1,11 +1,9 @@
+import { Heading, Tabs } from '@navikt/ds-react';
 import * as React from 'react';
-import { useState } from 'react';
 import SyntaxHighlighter from 'react-syntax-highlighter';
 import { docco } from 'react-syntax-highlighter/dist/esm/styles/hljs';
 import Block from '@navikt/sif-common-core-ds/src/atoms/block/Block';
 import { hasValue } from '../../../src/validation/validationUtils';
-import { Heading, Tabs } from '@navikt/ds-react';
-import Tab from '@navikt/ds-react/esm/tabs/Tab';
 
 interface Props {
     title: string;
@@ -15,7 +13,7 @@ interface Props {
 }
 
 const ValidationPanel: React.FunctionComponent<Props> = ({ title, children, code }) => {
-    const [activePanel, setActivePanel] = useState<string>('example');
+    const [activePanel, setActivePanel] = React.useState<string>('example');
     return (
         <div style={{ marginBottom: '4rem' }}>
             <Heading level="3" size="medium">
@@ -29,8 +27,8 @@ const ValidationPanel: React.FunctionComponent<Props> = ({ title, children, code
                             setActivePanel(idx);
                         }}>
                         <Tabs.List>
-                            <Tab value="example" label="Example" />
-                            <Tab value="code" label="Code" />
+                            <Tabs.Tab value="example" label="Example" />
+                            <Tabs.Tab value="code" label="Code" />
                         </Tabs.List>
                         <Tabs.Panel value={'example'}>{children}</Tabs.Panel>
                         <Tabs.Panel value={'code'}>